Added ability to infer hash algorithm
When using the -c flag, if an explicit algorithm was not specified,
one is inferred from the length of the hash. Only common hash types
are allowed to be inferred because many of the guaranteed algorithms
in hashlib produce the same length hash.

Updated the command line help text to indicate this; also moved the
algorithm choices into the epilog to make it cleaner.